Ok: Continue working out ownership details of OwningOk struct
We have implemented two constructors for the OwningOk struct which
assume that the type past does not have an const ro cv constraints.
We have deleted the constructors that might take a const type.

May be we should also add a line that panics if someone tries to
create an OwningOk with a const or cv qualified type